Modified structure of Devices, Lights
I removed hasBrightness
Then I noticed that brightness isn't really a property of "Devices", but
"Lights", and so moved them. I left in that last "pass" because it
didn't seem to hurt anything and I'm not real clear on how inheritance
works in Python.
Also ID'd a bug I had introduced in deviceDecoder(dct). I tried to fix
it but am not sure how to test.